refactor(pipeline): shim compile.py and logical_mvp.py to new pipelines
compile.compile_squin_to_move now delegates to PhysicalPipeline;
logical_mvp.compile_squin_to_move delegates to LogicalPipeline.
Both shims preserve the existing public API signatures.

Update test_compile_api_split.py to monkeypatch the new pipeline classes
instead of the removed squin_to_move import.  Fix two test files that
were accidentally importing squin_to_move via compile.py's transitive
namespace (test_upstream.py, test_validation_squin_kernels.py).
