+++ python/tests/test_validation_squin_kernels.py
@@ -6,12 +6,12 @@ from tests._validation_squin_kernels import (
 
 from bloqade.lanes.analysis import atom
 from bloqade.lanes.arch.gemini.physical import get_arch_spec as get_physical_arch_spec
-from bloqade.lanes.compile import squin_to_move
 from bloqade.lanes.dialects import move
 from bloqade.lanes.heuristics.physical.layout import (
     PhysicalLayoutHeuristicGraphPartitionCenterOut,
 )
 from bloqade.lanes.heuristics.physical.placement import PhysicalPlacementStrategy
+from bloqade.lanes.upstream import squin_to_move
 from bloqade.lanes.validation.address import Validation
